requirements.txt
accelerate==0.19.0
anyio==3.6.2
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
fastapi==0.95.2
filelock==3.12.0
fsspec==2023.5.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.14.1
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
psutil==5.9.5
pydantic==1.10.8
pymongo==4.3.3
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.6.1
urllib3==2.0.2
uvicorn==0.22.0
